modules/locEventConsumer/requirements.txt
kafka-python
SQLAlchemy
aniso8601
attrs
Click
Flask
flask-accepts
flask-cors
Flask-RESTful
flask-restplus
Flask-Script
Flask-SQLAlchemy
GeoAlchemy2
itsdangerous
Jinja2
jsonschema
MarkupSafe
marshmallow
marshmallow-sqlalchemy
psycopg2-binary
pyrsistent
python-dateutil
pytz
six
shapely
Werkzeug
flask-restx
